pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slurmkit"
version = "0.1.0"
description = "CLI tools for managing and generating SLURM jobs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: System :: Clustering",
    "Topic :: Utilities",
]
keywords = ["slurm", "hpc", "job-management", "cluster", "batch"]

dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.0",
    "pandas>=1.3",
    "tabulate>=0.9",
    "requests>=2.31",
]

[project.optional-dependencies]
ui = [
    "rich>=13.7",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
]

[project.scripts]
slurmkit = "slurmkit.cli.main:main"

[project.urls]
Homepage = "https://github.com/Awni00/slurm-job-utils"
Repository = "https://github.com/Awni00/slurm-job-utils"
Issues = "https://github.com/Awni00/slurm-job-utils/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
